backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
sqlmodel
openai
yt-dlp
langchain>=0.1.0
langchain-openai
langchain-community
langchain-core
langgraph
pydantic>=2.0
httpx
